New install approach, misc fixes
- Persist on main Spotify install and %localappdata% for config
- DLL hijacking instead of injection
- Fix Liked Songs playlist bricking

BlockTheSpot also targets dpapi.dll for hijacking, so Soggfy won't be
compatible with it. dpapi is the easiest target because it doesn't need
export redirection (at least of yet).

netmsg.dll may be another possible target, but I'm not sure it's quite
reliable since it seems to be loaded exclusively for some error message
formatting stuff.